Implement clean cutting mechanism
Changes the way cuts are handled when transcoding in fifo mode. Instead of
having the player seek past the cut, it seeks only part of the way and the
rest of the cut is handled my discarding data. This avoids the video
corruption that can occur with h264 random access, and gives the fifo
synchronisation mechanism time to settle.