MythDisplayX11: Use accurate XRandR refresh rates
This fixes automatic frame-rate switching when using modern nvidia drivers.

Reapply parts of Mark Kendall's use of the more modern XRandR API from the
render branch, specifically the changes to GetVideoModes and
SwitchToVideoModes.

Also take on part of a later commit to filter out interlaced modes.

The code differs currently in picking the first connected output and not
checking the name against the current screen. This is because the
DisplayRes class has no m_screen property back at v29. This may need
adressing.