Revert commit 2738b98, but add robustness
Commit 2738b98 removed the functionality to download fanart and coverart
upon a manual search for metadata.
The reason was that it produces an endless loop inside the `Videodialog`
selection of the correct item.

I re-add this feature with more logging and - hopefully - adding
robustness.
We can now distinguish between a failure of the returned lookup data
and a hiccup in the implementation of the 'ReferenceCounter' classes.