Handle forced AV subtitles without needing a forced track to be selected
This improves the behaviour of mythfrontend when it automatically handles
forced subtirle tracks. The previous behaviour relied on automatically
selecting the forced track as a default for the user's preference. The new
behaviour attempts to find a non-forced track for the user's preference and
changes the subtitle toggle feature to be between forced and non-forced,
rather than on and off.