Fix failure to clear last forced subtitle in a sequence
Some types of subtitle, rather than being tagged with a duration, rely
on the next in sequence to clear them, specifically with an empty
subtitle at the end of the sequence. AddAVSubtitle has to handle forced
subtitles, causing them to be displayed even when subtitles are off. It
was detecting a subtitle being forced by looking for a "forced" flag
amongst the subtitle's rectangles. That failed to detect the case of an
empty subtitle. Hence the last non-empty forced subtitle was not being
cleared by the subsequent empty one. This commit changes the detection
condition, using instead a flag that denotes whether the subtitle is from
the currently selected forced track.